requirements.txt
python-hcl2==4.3.2
colorama==0.4.6
click==8.1.7
xgboost==2.0.3
scikit-learn==1.4.0
joblib==1.3.2
pandas==2.1.4
openai==1.12.0
python-dotenv==1.0.0